change of generations the inheritance of Pixies across generations now works!
The selection of pixies to be transferred is done at the end of the previous generation and their Genomes are stored in a vector nextMetagenome.
Apart from that the previous World is out of scope and gets deleted (this way, we can work on stack as often as possible).

The inheritPixie-Function works just like spawnPixie in the sense that it generates a new Pixie entity and places it on the grid. But this time it draws from the nextMetagenome-
vector when generating a new Brain-Template. DNA is mutated at this step.
It also checks if there are similar or identical genomes already existing in the population, and skips some of the topological
computation steps if so.